chore: diffs for all contracts

+++ src/JumpRateModel.sol
@@ -4,40 +4,38 @@ import "./InterestRateModel.sol";
 import "./SafeMath.sol";
 
 /**
- * @title Moonwell's JumpRateModel Contract
- * @author Moonwell
- */
+  * @title Moonwell's JumpRateModel Contract
+  * @author Moonwell
+  */
 contract JumpRateModel is InterestRateModel {
-    using SafeMath for uint256;
+    using SafeMath for uint;
 
-    event NewInterestParams(
-        uint256 baseRatePerTimestamp, uint256 multiplierPerTimestamp, uint256 jumpMultiplierPerTimestamp, uint256 kink
-    );
+    event NewInterestParams(uint baseRatePerTimestamp, uint multiplierPerTimestamp, uint jumpMultiplierPerTimestamp, uint kink);
 
     /**
      * @notice The approximate number of timestamps per year that is assumed by the interest rate model
      */
-    uint256 public constant timestampsPerYear = 31536000;
+    uint public constant timestampsPerYear = 31536000;
 
     /**
      * @notice The multiplier of utilization rate that gives the slope of the interest rate
      */
-    uint256 public multiplierPerTimestamp;
+    uint public multiplierPerTimestamp;
 
     /**
      * @notice The base interest rate which is the y-intercept when utilization rate is 0
      */
-    uint256 public baseRatePerTimestamp;
+    uint public baseRatePerTimestamp;
 
     /**
      * @notice The multiplierPerTimestamp after hitting a specified utilization point
      */
-    uint256 public jumpMultiplierPerTimestamp;
+    uint public jumpMultiplierPerTimestamp;
 
     /**
      * @notice The utilization point at which the jump multiplier is applied
      */
-    uint256 public kink;
+    uint public kink;
 
     /**
      * @notice Construct an interest rate model
@@ -46,9 +44,7 @@ contract JumpRateModel is InterestRateModel {
      * @param jumpMultiplierPerYear The multiplierPerTimestamp after hitting a specified utilization point
      * @param kink_ The utilization point at which the jump multiplier is applied
      */
-    constructor(uint256 baseRatePerYear, uint256 multiplierPerYear, uint256 jumpMultiplierPerYear, uint256 kink_)
-        public
-    {
+    constructor(uint baseRatePerYear, uint multiplierPerYear, uint jumpMultiplierPerYear, uint kink_) public {
         baseRatePerTimestamp = baseRatePerYear.mul(1e18).div(timestampsPerYear).div(1e18);
         multiplierPerTimestamp = multiplierPerYear.mul(1e18).div(timestampsPerYear).div(1e18);
         jumpMultiplierPerTimestamp = jumpMultiplierPerYear.mul(1e18).div(timestampsPerYear).div(1e18);
@@ -64,7 +60,7 @@ contract JumpRateModel is InterestRateModel {
      * @param reserves The amount of reserves in the market (currently unused)
      * @return The utilization rate as a mantissa between [0, 1e18]
      */
-    function utilizationRate(uint256 cash, uint256 borrows, uint256 reserves) public pure returns (uint256) {
+    function utilizationRate(uint cash, uint borrows, uint reserves) public pure returns (uint) {
         // Utilization rate is 0 when there are no borrows
         if (borrows == 0) {
             return 0;
@@ -80,14 +76,14 @@ contract JumpRateModel is InterestRateModel {
      * @param reserves The amount of reserves in the market
      * @return The borrow rate percentage per timestmp as a mantissa (scaled by 1e18)
      */
-    function getBorrowRate(uint256 cash, uint256 borrows, uint256 reserves) public view returns (uint256) {
-        uint256 util = utilizationRate(cash, borrows, reserves);
+    function getBorrowRate(uint cash, uint borrows, uint reserves) public view returns (uint) {
+        uint util = utilizationRate(cash, borrows, reserves);
 
         if (util <= kink) {
             return util.mul(multiplierPerTimestamp).div(1e18).add(baseRatePerTimestamp);
         } else {
-            uint256 normalRate = kink.mul(multiplierPerTimestamp).div(1e18).add(baseRatePerTimestamp);
-            uint256 excessUtil = util.sub(kink);
+            uint normalRate = kink.mul(multiplierPerTimestamp).div(1e18).add(baseRatePerTimestamp);
+            uint excessUtil = util.sub(kink);
             return excessUtil.mul(jumpMultiplierPerTimestamp).div(1e18).add(normalRate);
         }
     }
@@ -100,14 +96,10 @@ contract JumpRateModel is InterestRateModel {
      * @param reserveFactorMantissa The current reserve factor for the market
      * @return The supply rate percentage per timestmp as a mantissa (scaled by 1e18)
      */
-    function getSupplyRate(uint256 cash, uint256 borrows, uint256 reserves, uint256 reserveFactorMantissa)
-        public
-        view
-        returns (uint256)
-    {
-        uint256 oneMinusReserveFactor = uint256(1e18).sub(reserveFactorMantissa);
-        uint256 borrowRate = getBorrowRate(cash, borrows, reserves);
-        uint256 rateToPool = borrowRate.mul(oneMinusReserveFactor).div(1e18);
+    function getSupplyRate(uint cash, uint borrows, uint reserves, uint reserveFactorMantissa) public view returns (uint) {
+        uint oneMinusReserveFactor = uint(1e18).sub(reserveFactorMantissa);
+        uint borrowRate = getBorrowRate(cash, borrows, reserves);
+        uint rateToPool = borrowRate.mul(oneMinusReserveFactor).div(1e18);
         return utilizationRate(cash, borrows, reserves).mul(rateToPool).div(1e18);
     }
 }
